taskrun/resources: allow input and output storage resources
This fixes an issue where a Task that uses a secret for both input and
output resources leads to duplicate volumes in the generated Pod.

This happens because both `resources.AddInputResource()` and
`resources.AddOutputResources()` mount the necessary secrets to load
artifacts. However, when multiple resources use the same secret the secret
volume is added twice and leads to an invalid PodSpec.

This fix extends the work done in
https://github.com/tektoncd/pipeline/pull/1370

(cherry picked from commit 1bcbe43ada87da461220ad3180e73f9a63e218d6)